+++ client-app/src/main.jsx
@@ -7,11 +7,9 @@ import { Provider } from "react-redux";
 import store from "./state/Store.js";
 
 createRoot(document.getElementById("root")).render(
-  <StrictMode>
     <BrowserRouter>
       <Provider store={store}>
         <App />
       </Provider>
     </BrowserRouter>
-  </StrictMode>
 );
